wheeeeee

+++ src/controllers/NoteController.ts
@@ -1,39 +1,75 @@
-import NoteResponseInterface from "../interfaces/NoteResponse";
+import NoteResponse from "../interfaces/NoteRetrieveResponse";
 import NoteDeletionResponse from "../interfaces/NoteDeletionResponse";
 import NoteUpdateResponse from "../interfaces/NoteUpdatedResponse";
+import NotFound from "../responses/NotFound";
+import { ResponseCodes } from "../responses/ResponseCodes";
+import NoteCreationResponse from "../interfaces/NoteCreationResponse";
 
-async function RetriveNote(req, res) {
+async function RetriveNote(req, res, next) {
     req.data.note = req.data.note ? req.data.note : await req.app.database.notes.findOne({ nanoid: req.params.nanoid });
 
-    const response: NoteResponseInterface = {
+    if (!req.data.note) return next(new NotFound("Note to retrieve was not found."));
+
+    const response: NoteResponse = {
         error: false,
         note: req.data.note,
     };
-    return res.status(200).json(response);
+    return res.status(ResponseCodes.OK).json(response);
 }
 
-async function DeleteNote(req, res) {
-    req.app.database.notes.deleteOne({ nanoid: req.data.note._id });
+async function DeleteNote(req, res, next) {
+    if (!req.data.note) return next(new NotFound("Note to delete was not found."));
 
-    const response: NoteDeletionResponse = {
-        error: false,
-        message: "Note Queued for Deletion.",
-    };
-    return res.status(202).json(response);
+    try {
+        req.app.database.notes.deleteOne({ nanoid: req.data.note._id });
+
+        const response: NoteDeletionResponse = {
+            error: false,
+            message: "Note Queued for Deletion.",
+        };
+        return res.status(ResponseCodes.ACCEPTED).json(response);
+    } catch (e) {
+        return next(e);
+    }
 }
 
-async function UpdateNote(req, res) {
+async function UpdateNote(req, res, next) {
     req.data.note = req.data.note ? req.data.note : await req.app.database.notes.findOne({ nanoid: req.params.nanoid });
+
+    if (!req.data.note) return next(new NotFound("Note to update was not found."));
 
     if (req.body.name) req.data.note.name = req.body.name;
     if (req.body.description) req.data.note.description = req.body.description;
-
-    await req.data.note.save();
-    const response: NoteUpdateResponse = {
-        error: false,
-        message: "Note Successfully Updated",
-    };
-    return res.status(200).json(response);
+    try {
+        const NewNote = await req.data.note.save();
+        const response: NoteUpdateResponse = {
+            error: false,
+            note: NewNote.toSafeObject(),
+        };
+        return res.status(ResponseCodes.OK).json(response);
+    } catch (e) {
+        return next(e);
+    }
 }
 
-export { RetriveNote, DeleteNote, UpdateNote };
+async function CreateNote(req, res, next) {
+    const { name, description } = req.body;
+    try {
+        const TempNote = new req.app.database.note({
+            name: name,
+            description: description,
+        });
+
+        const CreatedNote = await TempNote.save();
+
+        const response: NoteCreationResponse = {
+            error: false,
+            note: CreatedNote.toSafeObject(),
+        };
+        return res.status(ResponseCodes.CREATED).json(response);
+    } catch (e) {
+        return next(e);
+    }
+}
+
+export { RetriveNote, DeleteNote, UpdateNote, CreateNote };
